fix(security): remediate audit findings in v2 code and unblock PostgreSQL
A ten-agent security and code-quality audit of the full codebase found defects
in the AutoMod/health code added in v2.0.0, plus a release-blocking PostgreSQL
failure that predates it. This commit fixes the ones I introduced and the one
that stops the server booting at all.

AutoMod (all introduced in v2.0.0)

* Cross-space message injection via unvalidated `alert_channel`. The action was
  validated only as "parses as i64", so a rule in a space you own could target
  any channel id on the instance — including another space's channel or a DM —
  and the dispatcher posted there. Worse, the alert was authored by the member
  who tripped the rule and written via the raw DB path, forging authorship and
  skipping membership, permission, timeout and length validation. Now: alert
  channels must resolve to a text channel in the rule's own space (checked at
  create, at update, and again at dispatch), alerts are authored by the
  Auto-Moderator system user, and rule/excerpt text is stripped of markup and
  length-bounded before it enters a message body.

* Regex CPU exhaustion. Pattern count was uncapped (keywords were capped at 200)
  and every pattern recompiled on every evaluation. Measured at ~1.2s of
  synchronous runtime-thread CPU per request. Added MAX_PATTERNS and a bounded
  process-wide compiled-pattern cache.

* Unicode evasion. `bad<ZWSP>word`, fullwidth and mathematical alphanumerics all
  defeated keyword matching outright. Matching now folds through NFKC and drops
  format and combining characters. Homoglyph substitution remains out of scope
  and is documented as such.

* Mention-flood counted only `<@id>`, so role mentions and @everyone/@here — the
  two canonical mass pings — scored zero against the rule that exists to stop
  them. Ids are now normalized numerically, so `<@1>` and `<@001>` are one target.

* Keyword/pattern caps were applied to the filtered list while evaluation
  iterated the raw one, so a payload of blanks bypassed them.

* Whole-word matching advanced by the needle length, skipping an overlapping
  occurrence that would have matched ("a-a" in "xa-a-a"). Advances by one char.

* Alerts were dropped for blocked messages — moderators were notified about
  everything except what was actually stopped.

* AutoMod never ran on message edits: post innocuous text, edit it to the banned
  content, and every blocking rule was bypassed permanently with no hit recorded.

* Multiple timeout actions applied last-wins, so a message tripping a 28-day rule
  and a newer 60-second rule ended in a 60-second timeout. Collapsed to max().

Database

* PostgreSQL could not start. `migrations_pg/20260708000001` applies JSONB
  operators to `user_settings.notifications`, which is declared TEXT, aborting
  migrations with `operator does not exist: text ->> unknown`. Cast explicitly.

* The v2 migration used `ADD COLUMN ... DEFAULT (datetime('now'))`, which SQLite
  rejects on a populated table. Reproduced. Constant default + UPDATE backfill,
  which also gives legacy rows the semantically correct value.

* The v2 PG migration declared TIMESTAMP columns in a table whose other
  timestamps are TEXT; the sqlx `Any` driver cannot decode native temporals, and
  one such column fails the whole row. Now TEXT, with timestamps bound explicitly.

* Booleans were bound as `i64::from(bool)`, which PostgreSQL rejects against a
  BOOLEAN column. Bind the bool directly. Fixed in automod and emojis, and the
  convention in CLAUDE.md — which is how the bug propagated — corrected along
  with the timestamp and ADD COLUMN rules.

Dependencies: axios 1.16.1 -> 1.18.1 and dompurify 3.4.11 -> 3.4.12 (both had
advisories and both are reachable). react-router deliberately left at 7.18.1:
its advisory is RSC-mode-only and this app uses BrowserRouter, so the fix would
be a major upgrade buying nothing.

RELEASE_NOTES records that PostgreSQL is unusable in the v2.0.0 tag itself.

Verified: cargo test --workspace green except two VP9 decode tests that fail
identically on a clean v1.0.0 checkout; clippy clean with -D warnings; fmt clean;
19 AutoMod engine tests including 6 new regressions for the evasion classes above.

+++ crates/paracord-db/src/emojis.rs
@@ -43,7 +43,7 @@ pub async fn create_emoji(
     .bind(guild_id)
     .bind(name)
     .bind(creator_id)
-    .bind(i64::from(animated))
+    .bind(animated)
     .fetch_one(pool)
     .await?;
     Ok(row)
